+++ docs/changelog.md
@@ -6,6 +6,7 @@
     - e.g. `Color.fromColor3(...)` instead of `Color.from("Color3", ...)`
     - e.g. `Color:toColor3()` instead of `Color:to("Color3")`
 - Added `Color.gray` as a shortcut for creating achromatic Colors
+- Added `Color.harmonies` for generating Color harmonies
 - Added links in the documentation for further reading on various color types
 
 ### Removed